working waypoint navigation
TO DO;;;;;;;;;;;;;;;;;;;;;;

In the navigation loop, before calling driveTo(dest), calculate the next waypoint
if the line is not straight and we calculate an intermediate point,
return and push that intermediate point as dest. that way when we get to the
intermediate point we will likely see the end point and can calculate a new
longer trajectory, plus keeps us off walls.

driving is set to true, and set to false when waypoint is reached, thus
the loop is already controled, just need the next waypoint